chore: split passport cli logic
